% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_anomaly_detection.R
\name{ga_check_anomaly}
\alias{ga_check_anomaly}
\title{A function to perform Anomaly Detection
This is based on Twitter's package}
\usage{
ga_check_anomaly(..., df = NULL, time_col = "date", value_col = NULL,
  title = NULL, anoms = 0.02, direction = direction,
  piecewise_median_period_wk = NULL, y_log = y_log, longterm = NULL)
}
\arguments{
\item{...}{extra values to be passed to AnomalyDetection::AnomalyDetectionTs.}

\item{df}{Data frame}

\item{time_col}{Column that has time data. We assume this is either `date`, `datehour`` from Google Analytics API}

\item{value_col}{Column that has value data}

\item{direction}{Direction of anomaly. Positive ("posi"), Negative ("neg") or "both".}

\item{longterm}{Increase anom detection efficacy for time series that are greater than a month.
This automatically becomes TRUE if the data is longer than 30 days.}

\item{time_unit}{Time unit for aggregation.}

\item{fun.aggregate}{Function to aggregate values.}

\item{e_value}{Whether expected values should be returned.}

\item{na_fill_type}{- Type of NA fill:
"previous" - Fill with previous non-NA value.
"value" - Fill with the value of na_fill_value.
"interpolate" - Linear interpolation.
"spline" - Spline interpolation.
NULL - Skip NA fill. Use this only when you know there is no NA.}

\item{na_fill_value}{- Value to fill NA when na_fill_type is "value"}
}
\description{
A function to perform Anomaly Detection
This is based on Twitter's package
}
