% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_time_normalised.R
\name{ga_time_normalised}
\alias{ga_time_normalised}
\title{Time normalised traffic}
\usage{
ga_time_normalised(viewId, first_day_pageviews_min = 2,
  total_unique_pageviews_cutoff = 500, days_live_range = 60,
  page_filter_regex = ".*", interactive_plot = TRUE)
}
\arguments{
\item{viewId}{The viewId to use}

\item{first_day_pageviews_min}{threshold for first day of content}

\item{total_unique_pageviews_cutoff}{threshold of minimum unique pageviews}

\item{days_live_range}{How many days to show}

\item{page_filter_regex}{Select which pages to appear}

\item{interactive_plot}{Whether to have a plotly or ggplot output}
}
\value{
A \link[googleAnalyticsR]{ga_model} object
}
\description{
Based on \url{http://www.dartistics.com/googleanalytics/int-time-normalized.html} by Tim Wilson
}
